refactor: Context 파일 entities model로 이동
