+++ src/App.tsx
@@ -2,10 +2,10 @@ import { BrowserRouter as Router } from "react-router-dom"
 import Header from "./widgets/layout/ui/Header.tsx"
 import Footer from "./widgets/layout/ui/Footer.tsx"
 import PostsManagerPage from "./pages/PostsManagerPage.tsx"
-import { UserContextProvider } from "./shared/model/UserContext.tsx"
-import { PostContextProvider } from "./shared/model/PostContext.tsx"
-import { PostParamsContextProvider } from "./shared/model/PostParamsContext.tsx"
-import { CommentContextProvider } from "./shared/model/CommentContext.tsx"
+import { UserContextProvider } from "./entities/user/model/UserContext.tsx"
+import { PostContextProvider } from "./entities/post/model/PostContext.tsx"
+import { PostParamsContextProvider } from "./entities/post/model/PostParamsContext.tsx"
+import { CommentContextProvider } from "./entities/comment/model/CommentContext.tsx"
 
 const App = () => {
   return (
