+++ src/features/comment-add/ui/CommentAddButton.tsx
@@ -1,5 +1,5 @@
 import { Plus } from "lucide-react"
-import { useCommentContext } from "../../../shared/model/CommentContext"
+import { useCommentContext } from "../../../entities/comment/model/CommentContext"
 import { Button } from "../../../shared/ui"
 
 interface Props {
